Clean up watch manager
Fixes #295
